Refactor code for drop in implementations
Change the structure of the code base to allow for future VRRP
implementations to be dropped in. Future implementations will need to be
derived from the abstract base classes and the self._conf_obj in
vyatta/vyatta_vrrp_vci.py updated to point to the new implementation.

Moved the keepalived code around to fit this paradigm, using a new
directory with the concrete classes for the keepalived implementation
under it. Currently this is one file config_file.py but will be expanded
to include the VRRP group representation