Move vif interfaces to their own top level list
Restructure the data given to the component such that any vif interfaces
underneath interfaces are moved to a top level list in the datastructure
to make processing easier.
With vifs in a top level list we can process all vrrp groups in the same
manner, rather then searching deeper into the data structure for each
interface to see if a vif vrrp group exists.