Refactor code to use logging instead of prints
Introduce logging to the code instead of print statements. Print
statements were clogging up the test runs so switched them out for
logging commands.

Minor change to how the update() method was called on the self._conf_obj
and added a few new abstract classes to help with the logging