Add tests for _get_group_start_index function
Add tests for a function that returns the indexes of every search term
in a string. This is useful for finding where each vrrp group starts,
will also be of use for finding sync group blocks.
Four tests added, one with no groups, one with one group, one test
with two groups, and one testing for the auto-generation string