Add tests for finding value of keepalived config
Added 8 tests that check for the value of a config line in a block of
config. The return value from the function - to be implemented - is a
Tuple containing the presence of the config True/False, and the value
which can be either:
	NOTFOUND - Config not found (Presence False)
	[None] - Present but no value (Presence True)
	<value> - The value for the key (Presence True)
The yang representation has [null] for keys that are present but have no
value. Keys such as rfc-compatibility fall under this bracket

This doesn't work for multiline config which will be dealt with later.