Implement _find_config_value function
Added the implementation for _find_config_value, this function looks
through a config block and returns one of three tuples.

1) (False, "NOTFOUND") - Key not in config
2) (True, [None]) - Key in config without value
3) (True, <value>) - Key in config and has a value

Format 2 is for config like "rfc-compatibility" who's presence changes the
functionality without setting anything, [None] is used because the yang
representation uses [null] for these config nodes.
Format 3 is for config like "priority 254", a key and a value where the
value is returned to the caller.