Add virtualenv setup and information
Good practice for python development is to use a virtual environment and
install needed packages in this environment. This also makes it easier
to pass around/on boarding of new developers.
Added to the README explaining how to setup a virtualenv and install the
libraries to run the tests for the project.