Add tests to find an interface in the yang repr
Added tests that will test the codes ability to find and return an
interface in the yang representation. This is required because of the
format of the keepalived.conf file, each vrrp block is disparate and
contains the name of the interface the group is attached to. The yang
representation is different, having all the vrrp groups for an interface
hanging off the interface in a list.
These tests will test the implementation of finding an interface in the
yang data structure acts correctly