Implement method to find interfaces in yang repr
Complete implementation to find an interface in the yang representation
or create it if it doesn't exist.
This code will be used to make it easier to add the VRRP yang
representation. Finding exactly where the group should be added for an
easy append to the interfaces group list.

This method relies on python's pass by assignment. We get pass a list in
and so long as the list isn't assigned inside the method the method
works on the outer scope deceleration (basically a reference). We take
the list if it's empty we create the interface dictionary populating the
tagnode, and vrrp sections and return it. If the list isn't empty we
look through it for an entry that has the tagnode value set to the
interface name passed into the method if it's found we return this
reference and if it isn't found we create the entry append it and return
the reference. At each stage we look for vifs and deal with them.