Add code to convert vrrp config to yang
Add implementation to take the keepalived.conf format and convert it to the
yang json format.

The code is split into sections, one for single line config and the other
for config blocks.

The single line config is stored in a dictionary, the
key of which is the yang name and the value is the search term to
look for in the keepalived config block for the vrrp group. The value is
replaced by either 1) the value found in the conifg (key value line:
priority 100) 2) [None] (presence value: use_vmac) 3) NOTFOUND (config not
present). The dictionary is then filtered down to the pairs where the value
is not NOTFOUND.

The config block section looks for the search term and then the next
index of "}" and stores every line in between as a list.