Add test for class attribute instead of _object_setup
Pylint complained about instance attributes inside of a function that
wasn't __init__ as we can't use __init__ with VCI I've going to move the
keepalived config file to a class attribute that each instance will share.
To make sure that we are actually performing operations on the correct
implementation we check for the _conf_obj each time we call get, set, and
check.
This commit adds test for the type checking we're going to do