Change _object_setup to a type check
Instead of setting up a new instance every time we call set, get, check
move the _conf_obj to a class attribute that will be setup when the class
is imported. Then every time we call the VCI methods we check that the
type of _conf_obj is still the correct implementation type.
Because of pythons typing/class setup it's possible to modify the class
attribute for every instance