Implement check function for config validation
The only validation that can't be done in yang is determining if the hello
source config is an address on the box. Currently this is validated using
a perl script that takes the address and tries to bind a socket to it.
This is a pretty decent run time check because of things like DHCP, so I've
duplicated this mechanic in the check function.
We search the config for every hello-source-address entry and try to bind
to it. If it exists, all good close the socket, if it doesn't exist throw
the error and the check fails.
The error might be improved and the actual check infrastructure paradigm
might change but just now this is the best option