Add implementation of VCI State get call
Add implementation of the state object that will query the keepalived dbus
VRRP Instance object for the object's state. A little bit of data cleaning
is done but it's minimal.

I did attempt to use the asyncio commands but it just didn't work with the
synchronous code. I'll need to come back to it later with async to see how
it works properly.