Add RPC rfc-intf-map implmentation
Add the python code to send the dbus query to find the recv interface and
register it with the VCI infra.
Changed some variable names to make the dbus objects clearer. vrrp_proxy
became vrrp_group_proxy for when we're acting on a particular group and
used vrrp_process_proxy for when we're querying the process/all of the
groups e.g. for a search to find an interface