Rename find_recv_intf to rfc_intf_map and return vrid as well
To actually match the functionality from the previous implmentation the
DBus call needs to return the interface name and the group id. This is the
uniqueness characteristic.
Updated the tests to look for two values in the returned call and in the
JSON returned.
Updated the yang rpc definition to include the group number and changed the
interface to a string so an empty string could be returned for non-rfc
groups.
Changed the implementation to pull out two values