Move rfc mapping code into ProcessControl
As the rfc check is done completely on the process rather than an
individual group moved to it into ProcessControl class.
Moved the rpc call into the Config object as well for the call.
Updated the tests for this as well.
Included a few more bits and pieces for the __init__() change that I forgot
in a previous commit