Convert vrrp_group_connection to a class
Some changes with the __init__ means it makes more sense for this file to
actually be a class. When we get to emitting signals, and running scripts
we need to know a bit more information (like what the interface, group, and
af) that the pydbus library doesn't provide us.
Moved to a class and changed the signatures of the functions that act on it