Implement summary and detail vrrp show commands
Add new file vyatta/show_vrrp_cmds.py that will implement the guts of the
show vrrp commands.
Each function expects a dictionary of a similar format to the state
information received from the VCI infra. This is so we can in the future
receive the information directly from the process if needed (there is some
debate on which is better tmp file or direct data).
This means there will need to be a conversion step for the keepalived.data
file into the expected format.