Switch to f-strings and in line type checking
With the latest release of the Router the base OS includes python3.7 so
moved the code to use some new features.
f-strings are faster than .fromat() calls so replaced as much of them as
possible.
Type hinting can now be done inline rather than in a comment so replaced
these and added further hinting where it wasn't done before.