Implement script to return VRRP state
Added script/sbin/vyatta-vrrp-start.py which will be used by yang to
fetch the state of VRRP groups. This is a python2 script that will call
DBus methods on a VRRP object and marshal the data into a yang friendly
format. This is useful for gettree calls from the CLI and for netconf
sessions to find similar information as "show vrrp" displays on the CLI.

This is an intermediary step to implement a feature request, it will be
replaced in the near future by a python 3 VCI system.

Closes:VRVDR-48270