Bugfix/use valid characters in vif DBus calls
When trying to refresh mac entries of VRRP groups on a VIF interface
using the garp RPC/template script the call would error out. This is due
to the character "." not being valid for a dbus path.
Added code to replace the first "." in an interface name with "_" which
is a valid character. Applied this change in both the template script
and the RPC python script

There should only be one "." in an interface.

Closes:VRVDR-49164