Bugfix/Add error checking to configd call
The vrrp state script can crash if the config tree_get_dict() errors out
because of some other problem in the get calls. This can happen with a
get call on vifs, but I can not reproduce on KVMs. The regressions seem
to hit this problem regularly.
Added a try/catch block around the call to configd.Client and
tree_get_dict. Will return an empty dictionary if the call fails.

Closes:VRVDR-50393