Bugfix/Don't get state for invalid paths
If an interface existed but was not configured then the vrrp state
script would error out with an unhelpful message of unexpected element
is present.
The cause of this was attempting to call a tree get on a config node
that doesn't exist.
Stopped this error problem by checking if the full path is valid using
node_exists on each element of the path and silently returning if any
node in the path doesn't exist.

Closes:VRVDR-50842