Change template scripts to use new scripts
Change how op mode commands are called, use the new python scripts
instead of the Perl scripts. This will be moved to yang later on testing
the functionality with this change.
For the template scripts the output needs to be printed, this shouldn't
be the case for the YANG commands.
Removed debugging print statements that had been left in the op mode
script