Sanitise vif names for DBus paths
Replace . in an interface's name with _ to make valid DBus paths.
Updated the type of af_family, it was str and should've been int.
Changed the way rfc_intf is defined because the interface reported by
Keepalived contains . and not _ for vif interfaces. Now the variable is set
with what's in the XmitIntf DBus value and if once sanitised it matches
the stored interface name the rfc_intf variable is set to the empty string