Make finding initial interface in YANG more pythonic
In Python for loops you can include an else clause, this clause is only
called if the loop terminates normally - not if a break clause is reached.
When we're looking for the interface in the YANG we break when we find it.
So if it's not found the loop completes and we know we need to create a
python dictionary for this interface entry.

Modified how normal interfaces and vif interfaces are found the YANG
representation to use this mechanism