Bugfix/Change how check warnings/errors are reported
I'd made a mistake trying to report validation warnings just trying to
print the warning. There is a new method for doing this under VCI infra
by reporting an error to the system.

Added a new constant for the VRRP namespace rather that using the specific
VRRP yang node name
Changed the validation errors to throw the error
Updated tests for the new method
Changed doc strings to reflect new values