Change prints to log warning and make VMWare non-fatal
All regressions are run on ESXi VMWare having rfc mode commits fail on this
platform causes issues. The CLI warnings from the legacy system are an
unintended side effect that we want to move away from.
Changed the error for RFC mode on VMWare to be a logging warning
Changed print statement for preempt-delay and preempt to logging warning
Changed print statement for length of RFC interface to logging warning