Change find_config_value to return an enum instead of a Tuple
One of the review suggestions was to modify the tuple returned into an
enum.
The way YANG reports a presence leaf means three values are required.
 1) Not configured
 2) Configured, but YANG leaf is presence only
 3) Configured, and the YANG lead has a value

The values for the first two are always the same NOTFOUND (1) and [None]
(2) and are easily converted into Enums. The third value is not so easy to
convert into an enum value with the other two. The enum needs to be defined
together statically, which doesn't work for the third type listed.
To get round this and keep the processing the same in the caller for each
enum type I've grouped the first two into an enum and then created a new
enum class to return the configured value for the third type.