Address PR comments
Move test import out of fixture.

Replace multiline delimiter \ with ()

Instead of using \ for a statement covering multiple lines use () as a
better practice.
Updated if statements and assignments.
Haven't changed lines that are to get around PEP8 line lengths.

Warn is deprecated, use warning instead

Missed a magic string replaced it with it's constant value.

Change is_local_address to return a bool

Instead of having the caller deal with the OSError when the IP address
isn't configured locally catch it inside the function and return False
return True when the IP address if configured locally.
Updated tests to deal with new format

Change start_delay to an int instead of a string

Remove .keys call was unnecessary

Updated TODO with JIRA ID to track follow up task

Fixed format of is_rfc_compat_configured no longer needed separate lines
for PEP8 and simplify it now that the warning is sent to the log instead of
printing to the console

Added comment explaining that while most interface types are of the format
<namespace>:<intf_type> vif interfaces are not, so we need to use find()
instead of index() for semi-colons to split on.

Updated variable signatures to use the new Enum of find_config_value()
introduced TYPE_CHECKING for show_vrrp_cmds to avoid PEP8 errors on the
Enum type not being found.

auth_type_value was already a string so changed how lower() was called on
it

Improved readability of track-interfaces/track interfaces config using
get() calls

Updated how imports for local code works now packaging has been sorted

Update name for getting IP version to get_ip_version instead of
what_ip_version and changed it to a single line removing the variable

Remove unused logging import for config_file

Rename and retype trying to find the interface list to insert values into
interface_list was typed as a Dict renamed to interface_types and added
another variable called interface_list that was is the actual insertion
list

Remove triple quote strings and use \n and double quotes for the
Keepalived config file format