Testing refactoring
A lot of the tests did the same thing with different data so moved a lot of
those changes out into fixtures and made better use of the
pytest.mark.parametrize annotation to stop repetition of code

Tried to make the tests more consistent with asserts changed to be of the
form:
	assert <call> == expected

With expected defined just before the assert, I feel this reads better.

Also updated the conftest.py file to have consistent white spacing
If a returned value was braces, square brackets, or parenthesis these have
been moved on to the same line as the return and the rest of the python
moved to have the correct indentation

Added some tests that were missing, sometime the UT report different
coverage stats between runs. I think this is mostly due to the different
interface types. I need to investigate more