Rewrite find_config_value to raise error when config not found
As part of the PR concerns were raised over having an anonymous Enum
floating around on the stack and potentially being a source of a memory
leak.
Changed the code to either return:
	A string (configured with a value)
	A list containing only None (configured as presence)
	Raise a ValueError if the the config string isn't present
The callers now deal with the ValueError.
Updated callers of find_config_value to deal with the new format.
Updated tests that called find_config_value, split out the tests that now
error into a separate test