Make use of pythons try/except/else control flow
Where it makes sense (usually when a value was set to an empty list) use
the pythonic try/except/else control flow.
In places where we know if the key is found in the config it makes sense to
use this control flow rather than setting a variable to the empty list and
then immediately checking if that variable is of a type list.
Instead when the values are found drop immediately into parsing the keys
value.

This means some except branches only contain pass