Write extended tracking for switch interfaces to keepalived.conf
Added code to write the path monitor and route tracking to the Keepalived
config file for switch interface.
Switch interfaces were a new type of interface that VRRP was enabled on
while this rewrite was in progress.
Added tests for these interface types and improved the coverage for
converting between the two formats