Extend YANG pattern to allow switch VRRP intfs
There is currently a restriction in the "show interfaces vrrp" YANG to
only allow VRRP RFC interfaces of the format dpNvrrpN.
With the release of 2005 these generated VRRP interfaces can also take
the form swNvrrpN if RFC mode is used on a switch interface.
The YANG pattern prevents this new format of interface name being used
to filter "show interfaces vrrp".
This change set extends the pattern to allow switch generated
interfaces.

Closes:VRVDR-51497