Bugfix/Adjust VIF name in VCI notification
When using a VRRP group on a virtual interface the instances name is
generated as vyatta-<intf>_<vif_num>-<vrid> (vyatta-dp0p1s1_10-42) but
<intf>_<vif_num> is not the regular nomenclature <intf>.<vif_num> is.
The generated instance name should adjust in the notification instead of
having clients adjust it after the fact.

Added code to check for "_" in a group's instance name and replace it
with "." if it does.

Closes:VRVDR-52466