Bugfix/Report undef states as INIT
Reporting a state that wasn't INIT, FAULT, BACKUP, MASTER, or UNKNOWN
would cause issues with show commands when start-delay is used after a
group was in MASTER state and then "restart vrrp" had been called.
During the start-delay window the state would be reported as GOTOMASTER
which doesn't match the YANG.

Updated the State VCI class to return INIT if an unrecognised state is
returned.

Closes:VRVDR-53834