Bugfix/Allow tracking with syncgroups - no weights
There are many issues with syncgroups and tracking by the expected
behaviour is for any weighting to be ignored and still be able to use
the state tracking.
Without the additional config value sync_group_tracking_weight added to
the syncgroups then the tracking is completely ignored.
Introduced this config for all syncgroups and updated the tests for
syncgroups