Feature/Configure VRF aware RFC interfaces
Added new YANG leaf for VRF awareness, as this only makes sense with
rfc-compatibility it must be an empty node rather than boolean as
the XPATH1.1 expressions can not express an if statement.
Added code to the show section for the new config and added code to the
VRRP group class that will add the VRF option to the keepalived.conf
file.
Extended the test data for this new feature

Closes:VRVDR-39513