Bugfix/Update show vrrp sync for new data format
Keepalivedv2.2.0 changed the format of the data file for syncgroups,
which causes "show vrrp sync" to fail and subsequently broke connsync.
This was missed because the test data wasn't updated in the test.
Updated the tests to use the new keepalived.data format for sync groups.
Fixed the resulting issues in the code base from this format change.
Ignore every part of the sync group that doesn't contain "vyatta" only
the instances contain this.
Interfaces are now part of the data file after syncgroups so don't grab
everything after the syncgroup notifier.

Closes:VRVDR-56251